backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
dj-database-url==2.1.0
mysqlclient==2.2.4; sys_platform != "win32"
PyMySQL==1.1.0
Pillow>=10.2.0,<13.0.0
openpyxl==3.1.2
reportlab==4.1.0
requests==2.31.0
httpx==0.27.0
APScheduler==3.10.4
pytz==2024.1
